requirements.txt
# Core dependencies
aiogram>=3.0.0
sqlalchemy>=2.0.0
alembic>=1.0.0
asyncpg>=0.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI & API
openai>=1.0.0

# Web framework & server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# Database & Cache
redis>=5.0.0
psycopg2-binary>=2.9.0

# Utils
setuptools>=68.0.0
greenlet>=2.0.0
pytz>=2023.0.0
requests>=2.31.0

# Production extras
sentry-sdk>=1.0.0